+++ CHANGELOG.md
@@ -14,6 +14,10 @@
   is collected.
 * Add a `branchHits` field to `HitMap`.
 * Add support for scraping the service URI from the new Dart VM service message.
+* Replace `--packages` flag with `--package`, which takes the package's root
+  directory, instead of the .package file.
+* Deprecate the packagesPath parameter and add packagePath instead, in
+  `HitMap.parseJson`, `HitMap.parseFiles`, `createHitmap`, and `parseCoverage`.
 
 ## 1.1.0 - 2022-1-18